Fix brief flash of stale room data when joining a new room
